test: Define declarations
The testing includes the valid forms using the generic as well as required keys
'id' and 'version', both in singular and multiple-seqential form.

Furthermore, the testing also includes negative assertions for incorrect
identifiers as well as values. Although identifiers will be tested with its own
tests, it is important that it is also tested within usage context.