feat: TreeSitter visitors finished.
Currently, the TreeSitter compiles the coreLang and produces the exact same output as the ANTLR version.

The biggest change is that instead of using cursor.goto_next_sibling(), an auxiliary method was created to avoid comment nodes which still appear in the AST.

The grammar was updated to allow for multiple tags.